Triangles Module

Shiny module containing `mod_triangles_ui` and `mod_triangles_server`, respectively.

This module renders a user interface for displaying and analyzing actuarial loss data in the form of loss development triangles.

A shiny Module.

Usage

mod_triangles_ui(id, loss_data = loss_data_all)

mod_triangles_server(id, loss_data, selected_eval)

Arguments

id

ID associated with UI counterpart

loss_data

loss data

selected_eval

selected evaluation date