Create Triangle Bundle

Create a "bundle" of triangle related items from an input loss dataset. The function returns a list of class "triangle_bundle" and an attribute describing which metric the bundle describes (i.e. paid, reported, counts, etc.).

The resulting list contains:

- aggregated data filtered for ages of maturity from the `age_increment` argument

- triangle data derived off the aggregated data

- the actual spread out triangle

- age_to_age data

- the age_to_age spread out triangle

- averages (currently only straight and weighted)

- initial selections for the LDF's and derived CDF's.

Usage

create_triangle_bundle(
  loss_data,
  age_increment = 12,
  origin_col = "accident_year",
  age_col = "devt",
  value_col = "paid"
)

Arguments

loss_data

initial aggregated loss data as a `data.frame`

age_increment

increment in months between subsequent maturity periods

origin_col, age_col, value_col

column names as strings

Value

list of class "triangle_bundle" with an added attribute describing which metric the bundle describes (i.e. paid, reported, counts, etc.)

Examples

library(lossdevtapp)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# create default paid triangle bundle
tri_paid_bundle <- create_triangle_bundle(loss_data_all)

# check out the structure
str(tri_paid_bundle)
#> List of 7
#>  $ aggregate_data     : tibble [105 × 6] (S3: tbl_df/tbl/data.frame)
#>   ..$ accident_year: num [1:105] 2006 2006 2006 2006 2006 ...
#>   ..$ devt         : num [1:105] 12 24 36 48 60 72 84 96 108 120 ...
#>   ..$ paid         : num [1:105] 11146371 25763467 33902428 36367500 37392655 ...
#>   ..$ reported     : num [1:105] 24684326 34467713 36588672 37393391 37666758 ...
#>   ..$ n_claims     : int [1:105] 653 678 678 678 678 678 678 678 678 678 ...
#>   ..$ case         : num [1:105] 13537955 8704246 2686244 1025891 274103 ...
#>  $ triangle_data      : dev_tri [105 × 3] (S3: dev_tri/tbl_df/tbl/data.frame)
#>   ..$ origin: num [1:105] 2006 2006 2006 2006 2006 ...
#>   ..$ age   : num [1:105] 12 24 36 48 60 72 84 96 108 120 ...
#>   ..$ value : num [1:105] 11146371 25763467 33902428 36367500 37392655 ...
#>   ..- attr(*, "col_specs")=List of 3
#>   .. ..$ origin_col: chr "origin"
#>   .. ..$ age_col   : chr "age"
#>   .. ..$ value_col : chr "value"
#>  $ triangle           : tibble [14 × 15] (S3: tbl_df/tbl/data.frame)
#>   ..$ origin: num [1:14] 2006 2007 2008 2009 2010 ...
#>   ..$ 12    : num [1:14] 11146371 10545721 11597931 11122196 10868557 ...
#>   ..$ 24    : num [1:14] 25763467 23976020 27109408 24587682 24884104 ...
#>   ..$ 36    : num [1:14] 33902428 32018479 33165918 29929075 32053872 ...
#>   ..$ 48    : num [1:14] 36367500 35519541 35203610 31459144 35636080 ...
#>   ..$ 60    : num [1:14] 37392655 36110056 35922598 32470316 36173041 ...
#>   ..$ 72    : num [1:14] 37788200 36371766 35998095 33556727 36410318 ...
#>   ..$ 84    : num [1:14] 37811665 36434500 36039411 33647464 36517661 ...
#>   ..$ 96    : num [1:14] 37811665 36455980 36039411 33656817 36517661 ...
#>   ..$ 108   : num [1:14] 37811665 36455980 36039411 33664112 36517661 ...
#>   ..$ 120   : num [1:14] 37811665 36455980 36039411 33664112 36517661 ...
#>   ..$ 132   : num [1:14] 37811665 36455980 36039411 33664112 NA ...
#>   ..$ 144   : num [1:14] 37811665 36455980 36039411 NA NA ...
#>   ..$ 156   : num [1:14] 37811665 36455980 NA NA NA ...
#>   ..$ 168   : num [1:14] 37811665 NA NA NA NA ...
#>  $ age_to_age_data    : ata [91 × 3] (S3: ata/tbl_df/tbl/data.frame)
#>   ..$ origin: num [1:91] 2006 2006 2006 2006 2006 ...
#>   ..$ age   : num [1:91] 12 24 36 48 60 72 84 96 108 120 ...
#>   ..$ value : num [1:91] 2.31 1.32 1.07 1.03 1.01 ...
#>  $ age_to_age_triangle: tibble [13 × 14] (S3: tbl_df/tbl/data.frame)
#>   ..$ origin: num [1:13] 2006 2007 2008 2009 2010 ...
#>   ..$ 12    : num [1:13] 2.31 2.27 2.34 2.21 2.29 ...
#>   ..$ 24    : num [1:13] 1.32 1.34 1.22 1.22 1.29 ...
#>   ..$ 36    : num [1:13] 1.07 1.11 1.06 1.05 1.11 ...
#>   ..$ 48    : num [1:13] 1.03 1.02 1.02 1.03 1.02 ...
#>   ..$ 60    : num [1:13] 1.01 1.01 1 1.03 1.01 ...
#>   ..$ 72    : num [1:13] 1 1 1 1 1 ...
#>   ..$ 84    : num [1:13] 1 1 1 1 1 ...
#>   ..$ 96    : num [1:13] 1 1 1 1 1 ...
#>   ..$ 108   : num [1:13] 1 1 1 1 1 NA NA NA NA NA ...
#>   ..$ 120   : num [1:13] 1 1 1 1 NA NA NA NA NA NA ...
#>   ..$ 132   : num [1:13] 1 1 1 NA NA NA NA NA NA NA ...
#>   ..$ 144   : num [1:13] 1 1 NA NA NA NA NA NA NA NA ...
#>   ..$ 156   : num [1:13] 1 NA NA NA NA NA NA NA NA NA ...
#>  $ averages           : tibble [2 × 14] (S3: tbl_df/tbl/data.frame)
#>   ..$    : chr [1:2] "Straight Average:" "Weighted Average:"
#>   ..$ 12 : num [1:2] 2.32 2.32
#>   ..$ 24 : num [1:2] 1.27 1.27
#>   ..$ 36 : num [1:2] 1.08 1.08
#>   ..$ 48 : num [1:2] 1.02 1.02
#>   ..$ 60 : num [1:2] 1.01 1.01
#>   ..$ 72 : num [1:2] 1 1
#>   ..$ 84 : num [1:2] 1 1
#>   ..$ 96 : num [1:2] 1 1
#>   ..$ 108: num [1:2] 1 1
#>   ..$ 120: num [1:2] 1 1
#>   ..$ 132: num [1:2] 1 1
#>   ..$ 144: num [1:2] 1 1
#>   ..$ 156: num [1:2] 1 1
#>  $ selections         : tibble [2 × 14] (S3: tbl_df/tbl/data.frame)
#>   ..$    : chr [1:2] "Selected:" "CDF:"
#>   ..$ 12 : num [1:2] 2.32 3.29
#>   ..$ 24 : num [1:2] 1.27 1.42
#>   ..$ 36 : num [1:2] 1.08 1.12
#>   ..$ 48 : num [1:2] 1.02 1.03
#>   ..$ 60 : num [1:2] 1.01 1.01
#>   ..$ 72 : num [1:2] 1 1
#>   ..$ 84 : num [1:2] 1 1
#>   ..$ 96 : num [1:2] 1 1
#>   ..$ 108: num [1:2] 1 1
#>   ..$ 120: num [1:2] 1 1
#>   ..$ 132: num [1:2] 1 1
#>   ..$ 144: num [1:2] 1 1
#>   ..$ 156: num [1:2] 1 1
#>  - attr(*, "class")= chr [1:2] "tri_bundle" "list"
#>  - attr(*, "type")= chr "paid"

# derive a similar bundle for reported dollars and counts
tri_rept_bundle <- create_triangle_bundle(loss_data_all, value_col = "reported")
tri_cnts_bundle <- create_triangle_bundle(loss_data_all, value_col = "n_claims")