App Header

Shiny App Dashboard's Header Function

Functions that build the Shiny App's UI:

- `app_ui`: Main UI function

- `app_header`: wrapper around [shinydashboard::dashboardHeader()]

- `app_sidebar`: wrapper around [shinydashboard::dashboardSidebar()]

- `app_body`: wrapper around [shinydashboard::dashboardBody()]

Shiny App's User Interface Function.

Usage

app_header(title = "Loss Development", ...)

app_ui(request, ...)

Arguments

title

App title to be placed in the header, above the sidebar.

...

Arguments passed on to shinydashboardPlus::dashboardHeader

titleWidth

The width of the title area. This must either be a number which specifies the width in pixels, or a string that specifies the width in CSS units.

disable

If TRUE, don't display the header bar.

.list

An optional list containing items to put in the header. Same as the ... arguments, but in list format. This can be useful when working with programmatically generated items.

leftUi

Items that will appear on the left part of the navbar. Should be wrapped in a tagList.

controlbarIcon

Customize the trigger icon of the right sidebar.

fixed

Whether the navbar is fixed-top or not. FALSE by default.

request

Internal parameter for `shiny`.

Value

a [shinydashboard::dashboardHeader()]

The user interface definition, without modifications or side effects.

See also